LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/back-beforeunload-expected.txt



FAIL back() inside onbeforeunload assert_equals: expected 2 but got 1
